[PATCH] easily tweakable comm length
This change still keeps the comm length at 16, but allows easier patching
for local modifications, and also introduces a macro to use instead of
magic 16, where sizeof(comm) is not preferable to use.

Not able to use killall, pidof etc.  effectively, when long process names
are used for scripts.  Just changing the command length from 16 to 32
breaks a.out coredump logic.  Deamonise and get_task_comm helped in other
places in 2.6.10.

Signed-off-by: Prasanna Meda <pmeda@akamai.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>