[PATCH] PPC64 can do preempt debug too
This patch enables the DEBUG_PREEMPT config option for PPC64.  I have
this turned on on my desktop G5 and it isn't finding any problems.
(It did find one problem, in flush_tlb_pending(), that I have just
sent a patch for.)

BTW, do we really need to restrict which architectures the config
option is available on?

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>