[PATCH] USB: add driver for the Siemens ID Mouse fingerprint sensor
This patch adds a new usb-misc driver for the fingerprint sensor that
can be found in the Siemens ID Mouse USB. "cat /dev/usb/idmouseX"
yields a 225x288 greyscale PNM with the fingerprint information.

It's now in version 0.5, which uses memcpy() instead of snprintf()
and allows interruption of the image acquisition process, in case
it should get stuck.

It may be considered controversial that it outputs a PNM instead
of raw data, but I hold the opinion that the 15 bytes of header,
"P5 225 288 256 ", do not do any harm and allow the device to be
used in shell scripts or similar, too.

The setup packets are not described further, simply because I don"t
know anything about them myself. We captured them under Windows
using SnoopyPro.

Please include this into the main USB kernel tree - I think it has
by now been scrutinized and tested quite thoroughly.

Signed-off-by: Florian Echtler  <echtler@fs.tum.de>
Signed-off-by: Andreas Deresch <aderesch@fs.tum.de>