[PATCH] rename 'lock' to 'slock' in asm-i386/spinlock.h
This renames the x86 implementation of the spinlock_t's 'lock' field to
'slock', to protect against spinlock_t/rwlock_t type mismatches.

This way, if you use a spinlock where a rwlock is expected (or vice
versa), you'll get an obvious compile failure.

build- and boot-tested on x86 SMP+PREEMPT and SMP+!PREEMPT.

Signed-off-by: Ingo Molnar <mingo@elte.hu>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>