[PATCH] ppc32: update_process_times simplification
When the update_process_times call was moved out of do_timer for the UP case,
the replicator didn't track down the hiding and just added ifndef SMP.

This removes the ifdefs and the indirection of calling another file for one
function in a third file.

I also removed the per-cpu multipler and counter.  On some architectures (x86
and sparc64) these are set via /proc/profile setup_profiling_timer to increase
the timer interrupt by a factor over HZ.  However, this was not not
implemented for ppc (like several other architectures), so it was just
decrementing, testing, and setting per-cpu variables back to 1.

Signed-off-by: Milton Miller <miltonm@bga.com>
Signed-off-by: Tom Rini <trini@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>