[IA64] domain.c: eliminate warning when compiling CONFIG_NUMA=n
When compiling for non-NUMA variable "node" ended up not being used,
which solicits a warning from GCC.  Fix is to evluate cpu_to_node(i)
in place.  This has the effect of doing cpu_to_node(i) twice on NUMA,
but this is init code, so performance is not an issue (and even if
it were, you can just declare cpu_to_node() as being a pure function,
so the compiler can eliminate the second call).

Signed-off-by: David Mosberger-Tang <davidm@hpl.hp.com>
Signed-off-by: Tony Luck <tony.luck@intel.com>