kbuild: Introdude KBUILD_NOCMDDEP
When tossing around with different gcc compilers there is no way to tell kbuild
to ignore the new name of the compiler. The new option KBUILD_NOCMDDEP tell
kbuild not to check the commandline for changes.
This should be used with care because the resulting kernel may become inconsistent
if one part is build with 2.96, and another part build with 3.3.4.
So use only when you know what you are doing.

Syntax:
make KBUILD_NOCMDDEP=1

Original request for this feature came from hpa.

Signed-off-by: Sam Ravnborg <sam@ravnborg.org>