[PATCH] smc91x: allow RX of VLAN packets
This patch allows for VLAN packets to be received.  The initial patch
was from Andrew de Quincey.

While there, it also adds a test against an impossible hardware state
in theory but that happened in practice where the chip returns a
packet length of 0 which, once the status words have been substracted,
causes the transfer of data with a negative length.  Flaky hardware is
probably to blame here but better guard ourselves against that than
crashing the kernel.  David Brownell was the happy victim of such hw.

Signed-off-by: Nicolas Pitre <nico@cam.org>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>