[PATCH] Devices.txt, update with LANANA
Attached is diff for bringing devices.txt uptodate with lanana.

Please note: The devices.txt file in your tree will now be for 2.6+ kernels
only.  2.6 specific allocations will now be given out more freely, and some
of the stuff marked for obsolete for 2.6 has been removed.  I put a note in
the file to let people know its for 2.6+ kernels only.

I wanted to rename the new file to devices-2.6+.txt and then make a link
from the old devices.txt to this new file, but diffing it became too ugly.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>